% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette.R
\name{create_palette}
\alias{create_palette}
\title{Function creates a color palette from a named vector of hexadecimal string values.}
\usage{
create_palette(palette_path, palette_colors = NULL)
}
\arguments{
\item{palette_path}{A string that sets the file path for the .rda file.}

\item{palette_colors}{A named vector with a readable color name and its corresponding hexadecimal string value.}
}
\description{
The function creates a readable .rda file for persisting a list of hexadecimal strings for
defining a palette.
}
\author{
Rick Dean
}
