% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_scale.R
\name{get_color_scale}
\alias{get_color_scale}
\title{Create discrete/continuous color scaling for a ggplot2 aesthetic color variable.}
\usage{
get_color_scale(
  palette_path,
  scale_title = NULL,
  scale_limits = c(NA, NA),
  discrete = TRUE
)
}
\arguments{
\item{palette_path}{A string that sets the file path for a palette list.}

\item{scale_title}{A string that sets the tile for the scale's legend. By default the title comes from the
variable name assigned to the fill aesthetic.}

\item{scale_limits}{A numeric vector of length two that sets the limits of the non-discrete scale based on the variable
assigned to the fill aesthetic.}

\item{discrete}{A logical that is TRUE the function returns a discrete color scale. If FALSE, a gradient
color scale is returned.}
}
\value{
Returns a color scale object that can be attached to ggplot2 plot object for scaling the
color aesthetic variable.
}
\description{
The function can return either a color scale for a discrete variable or a gradient color
scale for a continuous numeric variable.
}
\author{
Rick Dean
}
