% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fill_scale.R
\name{get_fill_scale}
\alias{get_fill_scale}
\title{Create discrete/continuous fill scaling for a ggplot2 aesthetic fill variable.}
\usage{
get_fill_scale(
  palette_path,
  scale_title = NULL,
  scale_limits = c(NA, NA),
  discrete = TRUE
)
}
\arguments{
\item{palette_path}{A string that sets the file path for a palette list. Argument is required.}

\item{scale_title}{A string that sets the tile for the scale's legend. By default the title comes from the
variable name assigned to the fill aesthetic.}

\item{scale_limits}{A numeric vector of length two that sets the limits of the non-discrete scale based on the variable
assigned to the fill aesthetic.}

\item{discrete}{A logical that if TRUE the function returns a discrete fill scale. If FALSE, a gradient
fill scale is returned.}
}
\value{
Returns a fill scale object that can be attached to ggplot2 plot object for scaling the
fill aesthetic variable.
}
\description{
The function can return either a fill scale for a discrete variable or a gradient fill
scale for a continuous numeric variable.
}
\author{
Rick Dean
}
