% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_palette.R
\name{read_palette}
\alias{read_palette}
\title{Function reads a color palette and returns a vector of hexadecimal strings.}
\usage{
read_palette(palette_path = NULL)
}
\arguments{
\item{palette_path}{A string that sets the file path for the .rda file to read.}
}
\value{
Function returns a vector of colors valued as hexadecimal strings.
}
\description{
The function reads a readable .rda file created by
\code{RpalettePkg::create_palette()} and returns a list of hexadecimal strings
suitable for ggplot2 applications.
}
\author{
Rick Dean
}
